feat(app): introduce PassquitoClient
Introduces a new class `PassquitoClient` in
`src/utils/passquito-client.ts`, which provides the following features:
- registration of a new user
- registration of a new credential for an existing user
- authentication with a discoverable credential
- authentication of a specific user

They are the same functions that used to be provided by
`src/utils/passquito.ts`. As these functions all take an instance of
`CredentialApi` and feel tedious, I decided to wrap them in a class.

`refreshTokens` is excluded from `PassquitoClient` because it is just an
alias to `CredentialsApi.refreshTokens`. It no longer exists in
`src/utils/passquito.ts`.

Introduces a new Pinia store for the singleton instance of
`PassquitoClient`. Views access it by calling `usePassquitoClientStore`.
The store also provides the singleton instance of `CredentialsApi` for
convenience.
