+++ app/src/views/RegisterDevice.vue
@@ -4,12 +4,9 @@ import { onMounted, ref, watch } from 'vue';
 import { useRouter } from 'vue-router';
 
 import { useCredentialStore } from '../stores/credential';
-import { useCredentialsApiStore } from '../stores/credentials-api';
 import { usePasskeyCapabilityStore } from '../stores/passkey-capability';
-import {
-  doRegistrationCeremonyForVerifiedUser,
-  getErrorName,
-} from '../utils/passquito';
+import { usePassquitoClientStore } from '../stores/passquito-client';
+import { getErrorName } from '../utils/passquito';
 
 // router
 const router = useRouter();
@@ -22,8 +19,8 @@ onMounted(() => {
   passkeyCapabilityStore.askForCapabilities();
 });
 
-// credentials API access
-const credentialsApiStore = useCredentialsApiStore();
+// Passquito client
+const passquitoClientStore = usePassquitoClientStore();
 
 // credential
 const credentialStore = useCredentialStore();
@@ -73,16 +70,13 @@ const onSubmit = async () => {
   try {
     console.log('starting registration...');
     // TODO: deal with a token expiration
-    await doRegistrationCeremonyForVerifiedUser(
-      credentialsApiStore.api,
-      {
-        idToken,
-        userInfo: {
-          username: username.value,
-          displayName: displayName.value,
-        }
-      },
-    );
+    await passquitoClientStore.client.doRegistrationCeremonyForVerifiedUser({
+      idToken,
+      userInfo: {
+        username: username.value,
+        displayName: displayName.value,
+      }
+    });
     console.log('finished registration!');
   } catch (err) {
     if (getErrorName(err) === 'InvalidStateError') {
